i#1680 large pages: Fix more Windows build issues with 326a0d5.
Add two more casts, and move definition of os_page_size in win32/os.c
out of !NOT_DYNAMORIO_CORE_PROPER.

Review-URL: https://codereview.appspot.com/310440043

+++ api/samples/instrace_x86.c
@@ -312,7 +312,7 @@ code_cache_init(void)
     dr_insert_clean_call(drcontext, ilist, where, (void *)clean_call, false, 0);
     /* Encode the instructions into memory and clean up. */
     end = instrlist_encode(drcontext, ilist, code_cache, false);
-    DR_ASSERT((end - code_cache) < page_size);
+    DR_ASSERT((size_t)(end - code_cache) < page_size);
     instrlist_clear_and_destroy(drcontext, ilist);
     /* Set the memory as just +rx now. */
     dr_memory_protect(code_cache, page_size, DR_MEMPROT_READ | DR_MEMPROT_EXEC);
